C tests: split test_bds_exhaustive into h5 (core) and h10 (deep) binaries (closes #19)
The combined binary violated the tier contract: H=5 takes <1s but H=10 takes
~10 minutes, so labelling the whole binary "core" broke `ctest -L core < 5 min`.

Split into two binaries:
- test_bds_exhaustive_h5 (core tier, CORE_TIMEOUT): H=5 K=0 via XMSS-MT 20/4,
  32 sigs — catches BDS state machine bugs in seconds.
- test_bds_exhaustive_h10 (deep tier, EXHAUSTIVE_TIMEOUT): H=10 K=0/2/4 via
  XMSS-SHA2_10_256, 1024 sigs each — full deep validation.

Also fix ci-scheduled.yml step name: "H=5 K=0,2,4" was wrong (K=2 and K=4 are
invalid for H=5); corrected to "H=5 K=0 + H=10 K=0,2,4". The ctest -R regex
is updated from test_bds_exhaustive to test_bds_exhaustive_h to match both new
binaries.
